requirements.txt
google-cloud-storage
oauth2client
gitpython